+++ mat3x3.cpp
@@ -435,14 +435,14 @@ mat3x3 mat3x3_covariance(std::vector<vec3> points)
 
 	vec3 mean = make_vec3(0, 0, 0);
 
-	for (int i = 0; i < points.size(); i++)
+	for (size_t i = 0; i < points.size(); i++)
 	{
 		mean = vec3_add_vec3(mean, points[i]);
 	}
 
 	vec3_mult(&mean, 1 / (double)points.size());
 
-	for (int i = 0; i < points.size(); i++)
+	for (size_t i = 0; i < points.size(); i++)
 	{
 		points[i] = vec3_subtract_vec3(points[i], mean);
 	}
@@ -451,7 +451,7 @@ mat3x3 mat3x3_covariance(std::vector<vec3> points)
 	{
 		for (int i = 0; i < 3; i++)
 		{
-			for (int k = 0; k < points.size(); k++)
+			for (size_t k = 0; k < points.size(); k++)
 			{
 				double add = *(&points[k].x + i) * *(&points[k].x + j);
 				mat.vals[j * 3 + i] += add;
